backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
anthropic==0.39.0
httpx<0.28
langgraph==0.0.28
tavily-python==0.3.3
pydantic==2.6.1
